mangling: use meta programming for known-type substitutions
It’s better to have a single definition of known types which is then used in the mangler, demangler and remangler.

NFC
